🐛  - modal is popping up; buttons not working
The modal for New habit is working once again... however the Exit and Submit are not performing as intended. Some of the code that easily read state is no longer working. There is some logic for example that passes around boolean values to determine which buttons to show/hide. Why this was working I'm sure if it was correct or why now there's so much difficulty. Again I think it goes that to 'proper' state/prop passing. Just starting to get the gist of it but there's some sticking points still. I feel as though out dev build was throwing values around like crazy and it worked... now trying to go by best practices I'm stuck figuring our what's happening. It would explain to some degree how the Layout was working previously. I removed it thinking it was a useless wrapper... not I'm thinking it was what was responsible for passing things again.   #94